Add debit account and gift types to project mapping UI and schema
