+++ components/ui/button.tsx
@@ -20,6 +20,7 @@ const buttonVariants = cva(
       },
       size: {
         default: 'h-10 py-2 px-4',
+        xs: 'h-6 px-2 rounded-full text-xs',
         sm: 'h-9 px-3 rounded-md',
         lg: 'h-11 px-8 rounded-md',
       },
